fix: Add explicit long cast for CURLOPT_SSLVERSION to fix Fedora 43 build
Issue:
- Fedora 43 has stricter type checking in libcurl headers
- curl_easy_setopt expects long argument, but CURL_SSLVERSION_TLSv1_2 is enum
- Error: "curl_easy_setopt expects a long argument" with -Werror

Solution:
- Add explicit (long) cast to all CURL_SSLVERSION_TLSv1_2 usages
- Affected files (9 total):
  - src/user_experience/system_tray/system_tray.c
  - src/provider/lrclib/lrclib_provider.c (2 locations)
  - src/provider/itunes/itunes_artwork.c
  - src/translator/deepl/deepl_translator.c
  - src/translator/claude/claude_translator.c
  - src/translator/gemini/gemini_translator.c
  - src/translator/openai/openai_translator.c
  - src/translator/common/translator_common.c

This fixes Fedora 43 build failures while maintaining compatibility with Fedora 42.

Fixes COPR build: https://copr.fedorainfracloud.org/coprs/build/9937906

+++ src/provider/itunes/itunes_artwork.c
@@ -33,7 +33,7 @@ char* itunes_search_artwork(const char *artist, const char *album, const char *t
     }
 
     // Enforce TLS 1.2 or higher for security
-    curl_easy_setopt(curl, CURLOPT_SSLVERSION, CURL_SSLVERSION_TLSv1_2);
+    curl_easy_setopt(curl, CURLOPT_SSLVERSION, (long)CURL_SSLVERSION_TLSv1_2);
 
     // Sanitize track title to remove YouTube IDs and file extensions
     char *clean_track = sanitize_title(track);
